backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
numpy==1.26.4
Pillow==10.3.0
rasterio==1.3.10
scipy==1.13.0
imageio==2.34.1
gunicorn==21.2.0
